fix: resolve circular import errors in error handling modules
- Extract error constants (ErrorCategory, ErrorCode, ErrorSeverity) to separate constants.ts file
- Update error-handling/index.ts to import and re-export from constants.ts
- Fix circular imports in retry.ts and model-fallback.ts by importing constants directly
- Update process-jobs route to import ErrorCategory/ErrorSeverity from constants.ts
- Resolves Vercel build warnings and ensures clean module resolution
